IPC_HLE: Split up network code
This was way too monolithic.

I don't really have any other words for this except,
"Seriously, what the hell??".
