+++ Source/Core/Core/IPC_HLE/Net/MacHelpers.h
@@ -1,0 +1,15 @@
+// Copyright 2016 Dolphin Emulator Project
+// Licensed under GPLv2+
+// Refer to the license.txt file included.
+
+#pragma once
+
+#include "Common/CommonTypes.h"
+
+namespace MacHelpers
+{
+
+void SaveMacAddress(u8* mac);
+void GetMacAddress(u8* mac);
+
+} // namespace MacHelpers
