+++ Source/Core/Core/IPC_HLE/Net/WII_IPC_HLE_Device_net.cpp
@@ -7,20 +7,17 @@
 #include <cstdlib>
 #include <string>
 
-#include "Common/CommonPaths.h"
 #include "Common/FileUtil.h"
-#include "Common/NandPaths.h"
 #include "Common/Network.h"
-#include "Common/SettingsHandler.h"
 #include "Common/StringUtil.h"
 
 #include "Core/ConfigManager.h"
 #include "Core/Core.h"
-#include "Core/ec_wii.h"
-#include "Core/IPC_HLE/ICMP.h"
 #include "Core/IPC_HLE/WII_IPC_HLE_Device_es.h"
-#include "Core/IPC_HLE/WII_IPC_HLE_Device_net.h"
-#include "Core/IPC_HLE/WII_Socket.h"
+#include "Core/IPC_HLE/Net/ICMP.h"
+#include "Core/IPC_HLE/Net/MacHelpers.h"
+#include "Core/IPC_HLE/Net/WII_IPC_HLE_Device_net.h"
+#include "Core/IPC_HLE/Net/WII_Socket.h"
 
 #ifdef _WIN32
 #include <ws2tcpip.h>
@@ -56,509 +53,6 @@ typedef struct pollfd pollfd_t;
 #define UNSUPPORTED_WSAPOLL 0
 #endif
 
-// **********************************************************************************
-// Handle /dev/net/kd/request requests
-CWII_IPC_HLE_Device_net_kd_request::CWII_IPC_HLE_Device_net_kd_request(u32 _DeviceID, const std::string& _rDeviceName)
-	: IWII_IPC_HLE_Device(_DeviceID, _rDeviceName)
-{
-}
-
-CWII_IPC_HLE_Device_net_kd_request::~CWII_IPC_HLE_Device_net_kd_request()
-{
-	WiiSockMan::GetInstance().Clean();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_kd_request::Open(u32 _CommandAddress, u32 _Mode)
-{
-	INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: Open");
-	Memory::Write_U32(GetDeviceID(), _CommandAddress + 4);
-	m_Active = true;
-	return GetDefaultReply();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_kd_request::Close(u32 _CommandAddress, bool _bForce)
-{
-	INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: Close");
-	if (!_bForce)
-		Memory::Write_U32(0, _CommandAddress + 4);
-	m_Active = false;
-	return GetDefaultReply();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_kd_request::IOCtl(u32 _CommandAddress)
-{
-	u32 Parameter     = Memory::Read_U32(_CommandAddress + 0xC);
-	u32 BufferIn      = Memory::Read_U32(_CommandAddress + 0x10);
-	u32 BufferInSize  = Memory::Read_U32(_CommandAddress + 0x14);
-	u32 BufferOut     = Memory::Read_U32(_CommandAddress + 0x18);
-	u32 BufferOutSize = Memory::Read_U32(_CommandAddress + 0x1C);
-
-	u32 ReturnValue = 0;
-	switch (Parameter)
-	{
-	case IOCTL_NWC24_SUSPEND_SCHEDULAR:
-		// NWC24iResumeForCloseLib  from NWC24SuspendScheduler (Input: none, Output: 32 bytes)
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_SUSPEND_SCHEDULAR - NI");
-		Memory::Write_U32(0, BufferOut); // no error
-		break;
-
-	case IOCTL_NWC24_EXEC_TRY_SUSPEND_SCHEDULAR: // NWC24iResumeForCloseLib
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_EXEC_TRY_SUSPEND_SCHEDULAR - NI");
-		break;
-
-	case IOCTL_NWC24_EXEC_RESUME_SCHEDULAR: // NWC24iResumeForCloseLib
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_EXEC_RESUME_SCHEDULAR - NI");
-		Memory::Write_U32(0, BufferOut); // no error
-		break;
-
-	case IOCTL_NWC24_STARTUP_SOCKET: // NWC24iStartupSocket
-		Memory::Write_U32(0, BufferOut);
-		Memory::Write_U32(0, BufferOut + 4);
-		ReturnValue = 0;
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_STARTUP_SOCKET - NI");
-		break;
-
-	case IOCTL_NWC24_CLEANUP_SOCKET:
-		Memory::Memset(BufferOut, 0, BufferOutSize);
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_CLEANUP_SOCKET - NI");
-		break;
-
-	case IOCTL_NWC24_LOCK_SOCKET: // WiiMenu
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_LOCK_SOCKET - NI");
-		break;
-
-	case IOCTL_NWC24_UNLOCK_SOCKET:
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_UNLOCK_SOCKET - NI");
-		break;
-
-	case IOCTL_NWC24_REQUEST_REGISTER_USER_ID:
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_REQUEST_REGISTER_USER_ID");
-		Memory::Write_U32(0, BufferOut);
-		Memory::Write_U32(0, BufferOut + 4);
-		break;
-
-	case IOCTL_NWC24_REQUEST_GENERATED_USER_ID: // (Input: none, Output: 32 bytes)
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_REQUEST_GENERATED_USER_ID");
-		if (config.CreationStage() == nwc24_config_t::NWC24_IDCS_INITIAL)
-		{
-			std::string settings_Filename(Common::GetTitleDataPath(TITLEID_SYSMENU, Common::FROM_SESSION_ROOT) + WII_SETTING);
-			SettingsHandler gen;
-			std::string area, model;
-			bool _GotSettings = false;
-
-			if (File::Exists(settings_Filename))
-			{
-				File::IOFile settingsFileHandle(settings_Filename, "rb");
-				if (settingsFileHandle.ReadBytes((void*)gen.GetData(), SettingsHandler::SETTINGS_SIZE))
-				{
-					gen.Decrypt();
-					area = gen.GetValue("AREA");
-					model = gen.GetValue("MODEL");
-					_GotSettings = true;
-				}
-			}
-			if (_GotSettings)
-			{
-				u8 area_code = GetAreaCode(area);
-				u8 id_ctr = config.IdGen();
-				u8 hardware_model = GetHardwareModel(model);
-
-				EcWii &ec = EcWii::GetInstance();
-				u32 HollywoodID = ec.getNgId();
-				u64 UserID = 0;
-
-				s32 ret = NWC24MakeUserID(&UserID, HollywoodID, id_ctr, hardware_model, area_code);
-				config.SetId(UserID);
-				config.IncrementIdGen();
-				config.SetCreationStage(nwc24_config_t::NWC24_IDCS_GENERATED);
-				config.WriteConfig();
-
-				Memory::Write_U32(ret, BufferOut);
-			}
-			else
-			{
-				Memory::Write_U32(WC24_ERR_FATAL, BufferOut);
-			}
-		}
-		else if (config.CreationStage() == nwc24_config_t::NWC24_IDCS_GENERATED)
-		{
-			Memory::Write_U32(WC24_ERR_ID_GENERATED, BufferOut);
-		}
-		else if (config.CreationStage() == nwc24_config_t::NWC24_IDCS_REGISTERED)
-		{
-			Memory::Write_U32(WC24_ERR_ID_REGISTERED, BufferOut);
-		}
-		Memory::Write_U64(config.Id(), BufferOut + 4);
-		Memory::Write_U32(config.CreationStage(), BufferOut + 0xC);
-		break;
-
-	case IOCTL_NWC24_GET_SCHEDULAR_STAT:
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_GET_SCHEDULAR_STAT - NI");
-		break;
-
-	case IOCTL_NWC24_SAVE_MAIL_NOW:
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_SAVE_MAIL_NOW - NI");
-		break;
-
-	case IOCTL_NWC24_REQUEST_SHUTDOWN:
-		// if ya set the IOS version to a very high value this happens ...
-		INFO_LOG(WII_IPC_WC24, "NET_KD_REQ: IOCTL_NWC24_REQUEST_SHUTDOWN - NI");
-		break;
-
-	default:
-		INFO_LOG(WII_IPC_WC24, "/dev/net/kd/request::IOCtl request 0x%x (BufferIn: (%08x, %i), BufferOut: (%08x, %i)",
-			Parameter, BufferIn, BufferInSize, BufferOut, BufferOutSize);
-		break;
-	}
-
-	Memory::Write_U32(ReturnValue, _CommandAddress + 4);
-	return GetDefaultReply();
-}
-
-
-u8 CWII_IPC_HLE_Device_net_kd_request::GetAreaCode(const std::string& area) const
-{
-	static std::map<const std::string, u8> regions = {
-		{ "JPN", 0 }, { "USA", 1 }, { "EUR", 2 },
-		{ "AUS", 2 }, { "BRA", 1 }, { "TWN", 3 },
-		{ "ROC", 3 }, { "KOR", 4 }, { "HKG", 5 },
-		{ "ASI", 5 }, { "LTN", 1 }, { "SAF", 2 },
-		{ "CHN", 6 },
-	};
-
-	auto entryPos = regions.find(area);
-	if (entryPos != regions.end())
-		return entryPos->second;
-	else
-		return 7; // Unknown
-}
-
-u8 CWII_IPC_HLE_Device_net_kd_request::GetHardwareModel(const std::string& model) const
-{
-	static std::map<const std::string, u8> models = {
-		{ "RVL", MODEL_RVL },
-		{ "RVT", MODEL_RVT },
-		{ "RVV", MODEL_RVV },
-		{ "RVD", MODEL_RVD },
-	};
-
-	auto entryPos = models.find(model);
-	if (entryPos != models.end())
-		return entryPos->second;
-	else
-		return MODEL_ELSE;
-}
-
-static inline u8 u64_get_byte(u64 value, u8 shift)
-{
-	return (u8)(value >> (shift*8));
-}
-
-static inline u64 u64_insert_byte(u64 value, u8 shift, u8 byte)
-{
-	u64 mask = 0x00000000000000FFULL << (shift*8);
-	u64 inst = (u64)byte << (shift*8);
-	return (value & ~mask) | inst;
-}
-
-s32 CWII_IPC_HLE_Device_net_kd_request::NWC24MakeUserID(u64* nwc24_id, u32 hollywood_id, u16 id_ctr, u8 hardware_model, u8 area_code)
-{
-	const u8 table2[8]  = {0x1, 0x5, 0x0, 0x4, 0x2, 0x3, 0x6, 0x7};
-	const u8 table1[16] = {0x4, 0xB, 0x7, 0x9, 0xF, 0x1, 0xD, 0x3, 0xC, 0x2, 0x6, 0xE, 0x8, 0x0, 0xA, 0x5};
-
-	u64 mix_id = ((u64)area_code << 50) | ((u64)hardware_model << 47) | ((u64)hollywood_id << 15) | ((u64)id_ctr << 10);
-	u64 mix_id_copy1 = mix_id;
-
-	int ctr = 0;
-	for (ctr = 0; ctr <= 42; ctr++)
-	{
-		u64 value = mix_id >> (52 - ctr);
-		if (value & 1)
-		{
-			value = 0x0000000000000635ULL << (42-ctr);
-			mix_id ^= value;
-		}
-	}
-
-	mix_id = (mix_id_copy1 | (mix_id & 0xFFFFFFFFUL)) ^ 0x0000B3B3B3B3B3B3ULL;
-	mix_id = (mix_id >> 10) | ((mix_id & 0x3FF) << (11 + 32));
-
-	for (ctr = 0; ctr <= 5; ctr++)
-	{
-		u8 ret = u64_get_byte(mix_id, ctr);
-		u8 foobar = ((table1[(ret >> 4) & 0xF]) << 4) | (table1[ret & 0xF]);
-		mix_id = u64_insert_byte(mix_id, ctr, foobar & 0xff);
-	}
-	u64 mix_id_copy2 = mix_id;
-
-	for (ctr = 0; ctr <= 5; ctr++)
-	{
-		u8 ret = u64_get_byte(mix_id_copy2, ctr);
-		mix_id = u64_insert_byte(mix_id, table2[ctr], ret);
-	}
-
-	mix_id &= 0x001FFFFFFFFFFFFFULL;
-	mix_id = (mix_id << 1) | ((mix_id >> 52) & 1);
-
-	mix_id ^= 0x00005E5E5E5E5E5EULL;
-	mix_id &= 0x001FFFFFFFFFFFFFULL;
-
-	*nwc24_id = mix_id;
-
-	if (mix_id > 9999999999999999ULL)
-		return WC24_ERR_FATAL;
-
-	return WC24_OK;
-}
-
-static void SaveMacAddress(u8* mac)
-{
-	SConfig::GetInstance().m_WirelessMac = MacAddressToString(mac);
-	SConfig::GetInstance().SaveSettings();
-}
-
-static void GetMacAddress(u8* mac)
-{
-	// Parse MAC address from config, and generate a new one if it doesn't
-	// exist or can't be parsed.
-	std::string wireless_mac = SConfig::GetInstance().m_WirelessMac;
-
-	if (Core::g_want_determinism)
-		wireless_mac = "12:34:56:78:9a:bc";
-
-	if (!StringToMacAddress(wireless_mac, mac))
-	{
-		GenerateMacAddress(IOS, mac);
-		SaveMacAddress(mac);
-		if (!wireless_mac.empty())
-		{
-			ERROR_LOG(WII_IPC_NET, "The MAC provided (%s) is invalid. We have "\
-			                       "generated another one for you.",
-			                       MacAddressToString(mac).c_str());
-		}
-	}
-	INFO_LOG(WII_IPC_NET, "Using MAC address: %s", MacAddressToString(mac).c_str());
-}
-
-// **********************************************************************************
-// Handle /dev/net/ncd/manage requests
-CWII_IPC_HLE_Device_net_ncd_manage::CWII_IPC_HLE_Device_net_ncd_manage(u32 _DeviceID, const std::string& _rDeviceName)
-	: IWII_IPC_HLE_Device(_DeviceID, _rDeviceName)
-{
-}
-
-CWII_IPC_HLE_Device_net_ncd_manage::~CWII_IPC_HLE_Device_net_ncd_manage()
-{
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_ncd_manage::Open(u32 _CommandAddress, u32 _Mode)
-{
-	INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: Open");
-	Memory::Write_U32(GetDeviceID(), _CommandAddress + 4);
-	m_Active = true;
-	return GetDefaultReply();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_ncd_manage::Close(u32 _CommandAddress, bool _bForce)
-{
-	INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: Close");
-	if (!_bForce)
-		Memory::Write_U32(0, _CommandAddress + 4);
-	m_Active = false;
-	return GetDefaultReply();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_ncd_manage::IOCtlV(u32 _CommandAddress)
-{
-	u32 return_value = 0;
-	u32 common_result = 0;
-	u32 common_vector = 0;
-
-	SIOCtlVBuffer CommandBuffer(_CommandAddress);
-
-	switch (CommandBuffer.Parameter)
-	{
-	case IOCTLV_NCD_LOCKWIRELESSDRIVER:
-		break;
-
-	case IOCTLV_NCD_UNLOCKWIRELESSDRIVER:
-		//Memory::Read_U32(CommandBuffer.InBuffer.at(0).m_Address);
-		break;
-
-	case IOCTLV_NCD_GETCONFIG:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETCONFIG");
-		config.WriteToMem(CommandBuffer.PayloadBuffer.at(0).m_Address);
-		common_vector = 1;
-		break;
-
-	case IOCTLV_NCD_SETCONFIG:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_SETCONFIG");
-		config.ReadFromMem(CommandBuffer.InBuffer.at(0).m_Address);
-		break;
-
-	case IOCTLV_NCD_READCONFIG:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_READCONFIG");
-		config.ReadConfig();
-		config.WriteToMem(CommandBuffer.PayloadBuffer.at(0).m_Address);
-		common_vector = 1;
-		break;
-
-	case IOCTLV_NCD_WRITECONFIG:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_WRITECONFIG");
-		config.ReadFromMem(CommandBuffer.InBuffer.at(0).m_Address);
-		config.WriteConfig();
-		break;
-
-	case IOCTLV_NCD_GETLINKSTATUS:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETLINKSTATUS");
-		// Always connected
-		Memory::Write_U32(netcfg_connection_t::LINK_WIRED, CommandBuffer.PayloadBuffer.at(0).m_Address + 4);
-		break;
-
-	case IOCTLV_NCD_GETWIRELESSMACADDRESS:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETWIRELESSMACADDRESS");
-
-		u8 address[MAC_ADDRESS_SIZE];
-		GetMacAddress(address);
-		Memory::CopyToEmu(CommandBuffer.PayloadBuffer.at(1).m_Address, address, sizeof(address));
-		break;
-
-	default:
-		INFO_LOG(WII_IPC_NET, "NET_NCD_MANAGE IOCtlV: %#x", CommandBuffer.Parameter);
-		break;
-	}
-
-	Memory::Write_U32(common_result, CommandBuffer.PayloadBuffer.at(common_vector).m_Address);
-	if (common_vector == 1)
-	{
-		Memory::Write_U32(common_result, CommandBuffer.PayloadBuffer.at(common_vector).m_Address + 4);
-	}
-	Memory::Write_U32(return_value, _CommandAddress + 4);
-	return GetDefaultReply();
-}
-
-// **********************************************************************************
-// Handle /dev/net/wd/command requests
-CWII_IPC_HLE_Device_net_wd_command::CWII_IPC_HLE_Device_net_wd_command(u32 DeviceID, const std::string& DeviceName)
-	: IWII_IPC_HLE_Device(DeviceID, DeviceName)
-{
-}
-
-CWII_IPC_HLE_Device_net_wd_command::~CWII_IPC_HLE_Device_net_wd_command()
-{
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_wd_command::Open(u32 CommandAddress, u32 Mode)
-{
-	INFO_LOG(WII_IPC_NET, "NET_WD_COMMAND: Open");
-	Memory::Write_U32(GetDeviceID(), CommandAddress + 4);
-	m_Active = true;
-	return GetDefaultReply();
-}
-
-IPCCommandResult CWII_IPC_HLE_Device_net_wd_command::Close(u32 CommandAddress, bool Force)
-{
-	INFO_LOG(WII_IPC_NET, "NET_WD_COMMAND: Close");
-	if (!Force)
-		Memory::Write_U32(0, CommandAddress + 4);
-	m_Active = false;
-	return GetDefaultReply();
-}
-
-// This is just for debugging / playing around.
-// There really is no reason to implement wd unless we can bend it such that
-// we can talk to the DS.
-IPCCommandResult CWII_IPC_HLE_Device_net_wd_command::IOCtlV(u32 CommandAddress)
-{
-	u32 return_value = 0;
-
-	SIOCtlVBuffer CommandBuffer(CommandAddress);
-
-	switch (CommandBuffer.Parameter)
-	{
-	case IOCTLV_WD_SCAN:
-		{
-		// Gives parameters detailing type of scan and what to match
-		// XXX - unused
-		// ScanInfo *scan = (ScanInfo *)Memory::GetPointer(CommandBuffer.InBuffer.at(0).m_Address);
-
-		u16* results = (u16*)Memory::GetPointer(CommandBuffer.PayloadBuffer.at(0).m_Address);
-		// first u16 indicates number of BSSInfo following
-		results[0] = Common::swap16(1);
-
-		BSSInfo* bss = (BSSInfo*)&results[1];
-		memset(bss, 0, sizeof(BSSInfo));
-
-		bss->length = Common::swap16(sizeof(BSSInfo));
-		bss->rssi = Common::swap16(0xffff);
-
-		for (int i = 0; i < BSSID_SIZE; ++i)
-			bss->bssid[i] = i;
-
-		const char* ssid = "dolphin-emu";
-		strcpy((char*)bss->ssid, ssid);
-		bss->ssid_length = Common::swap16((u16)strlen(ssid));
-
-		bss->channel = Common::swap16(2);
-		}
-		break;
-
-	case IOCTLV_WD_GET_INFO:
-		{
-		Info* info = (Info*)Memory::GetPointer(CommandBuffer.PayloadBuffer.at(0).m_Address);
-		memset(info, 0, sizeof(Info));
-		// Probably used to disallow certain channels?
-		memcpy(info->country, "US", 2);
-		info->ntr_allowed_channels = Common::swap16(0xfffe);
-
-		u8 address[MAC_ADDRESS_SIZE];
-		GetMacAddress(address);
-		memcpy(info->mac, address, sizeof(info->mac));
-		}
-		break;
-
-	case IOCTLV_WD_GET_MODE:
-	case IOCTLV_WD_SET_LINKSTATE:
-	case IOCTLV_WD_GET_LINKSTATE:
-	case IOCTLV_WD_SET_CONFIG:
-	case IOCTLV_WD_GET_CONFIG:
-	case IOCTLV_WD_CHANGE_BEACON:
-	case IOCTLV_WD_DISASSOC:
-	case IOCTLV_WD_MP_SEND_FRAME:
-	case IOCTLV_WD_SEND_FRAME:
-	case IOCTLV_WD_CALL_WL:
-	case IOCTLV_WD_MEASURE_CHANNEL:
-	case IOCTLV_WD_GET_LASTERROR:
-	case IOCTLV_WD_CHANGE_GAMEINFO:
-	case IOCTLV_WD_CHANGE_VTSF:
-	case IOCTLV_WD_RECV_FRAME:
-	case IOCTLV_WD_RECV_NOTIFICATION:
-	default:
-		INFO_LOG(WII_IPC_NET, "NET_WD_COMMAND IOCtlV %#x in %i out %i",
-			CommandBuffer.Parameter, CommandBuffer.NumberInBuffer, CommandBuffer.NumberPayloadBuffer);
-		for (u32 i = 0; i < CommandBuffer.NumberInBuffer; ++i)
-		{
-			INFO_LOG(WII_IPC_NET, "in %i addr %x size %i", i,
-				CommandBuffer.InBuffer.at(i).m_Address, CommandBuffer.InBuffer.at(i).m_Size);
-			INFO_LOG(WII_IPC_NET, "%s",
-				ArrayToString(
-					Memory::GetPointer(CommandBuffer.InBuffer.at(i).m_Address),
-					CommandBuffer.InBuffer.at(i).m_Size).c_str()
-				);
-		}
-		for (u32 i = 0; i < CommandBuffer.NumberPayloadBuffer; ++i)
-		{
-			INFO_LOG(WII_IPC_NET, "out %i addr %x size %i", i,
-				CommandBuffer.PayloadBuffer.at(i).m_Address, CommandBuffer.PayloadBuffer.at(i).m_Size);
-		}
-		break;
-	}
-
-	Memory::Write_U32(return_value, CommandAddress + 4);
-	return GetDefaultReply();
-}
-
-// **********************************************************************************
 // Handle /dev/net/ip/top requests
 CWII_IPC_HLE_Device_net_ip_top::CWII_IPC_HLE_Device_net_ip_top(u32 _DeviceID, const std::string& _rDeviceName)
 	: IWII_IPC_HLE_Device(_DeviceID, _rDeviceName)
@@ -1330,7 +824,7 @@ IPCCommandResult CWII_IPC_HLE_Device_net_ip_top::IOCtlV(u32 CommandAddress)
 
 		case 0x1004: // mac address
 			u8 address[MAC_ADDRESS_SIZE];
-			GetMacAddress(address);
+			MacHelpers::GetMacAddress(address);
 			Memory::CopyToEmu(_BufferOut, address, sizeof(address));
 			break;
